Major changes with extra commands.
Commands now start with "Piv" form.
Added commands Pylint, Pep8, AutoPep8, Vulture, Pydoc.
Also added a command to directly run the file in the buffer.
The only main thing of the plugin is not executing buffer contents anymore, but its unobtrusive bindings to Python and its linting tools.
The plugin still does not require Vim to be compiled with +python, all the work is done from VimScript and shell commands.